requirements_training.txt
# Qwen3-VL Fine-tuning Dependencies
# Install: pip install -r requirements_training.txt

# Core
torch>=2.4.0
torchvision>=0.19.0
transformers>=4.50.0

# Training
deepspeed>=0.15.0
accelerate>=1.0.0
peft>=0.12.0

# Attention optimization
flash-attn>=2.5.0  # Requires CUDA

# Data processing
Pillow>=10.0.0
numpy>=1.24.0
scipy>=1.10.0

# Evaluation
scikit-learn>=1.3.0

# Logging
tensorboard>=2.15.0
wandb>=0.16.0

# Hyperspectral processing (already installed)
# spectral>=0.23.0

